requirement_raspi.txt
wheel==0.37.0
mfrc522==0.0.7
RPi.GPIO==0.7.0
spidev==3.5
